// prisma/generated/zod/inputTypeSchemas/EnergyCountOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';
import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const EnergyCountOrderByAggregateInputSchema: z.ZodType<Prisma.EnergyCountOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  name: z.lazy(() => SortOrderSchema).optional(),
  unit: z.lazy(() => SortOrderSchema).optional(),
  info: z.lazy(() => SortOrderSchema).optional(),
  color: z.lazy(() => SortOrderSchema).optional(),
  music: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default EnergyCountOrderByAggregateInputSchema;